refactor: avatar
